Support SSR and enable deployments
[Closes #1]

Added multithreading dynamic app-db capabilities to the original re-frame
code, as described in the issue. In this case, only backend-side behaviour is
changed, keeping current frontend behaviour untouched, since we do not need to
change anything at this point. Regarding the implementation, we basically make
sure we can store different app-db instances per thread, exposing a way of
cleaning-up a specific app-db instance, which is needed after SSR processs has
finished for a given http request. Then, at the time of registering and
processing the events in the queue, we take into account the specific app-db
instance that is related to each event, when we are at the backend side.

As described in the issue, we are specifying the artifact version in the
`project.clj` namespace, in order to simplify the CI process. Hence, we do not
need custom `release-tasks` options, but we keep the `lein-git-inject` plugin
settings to minimize conflicts when updating the fork from upstream.

For making easier the fork's update from upstream, we have added minimal
changes to the `README.md` file, adding a link where we describe with more
details the specific additions of this fork to the original re-frame.

Separated deployment settings to allow snapshot deployments to Clojars.

We have also removed the github actions' related links from the `README.md`
file, since we are not using them for this repo, as we are just going to
deploy the code without changing anything else related to the original re-
frame. The idea is to use custom CI file for executing tests and deployments
with Github Actions, disabling the original ones to avoid issues.